% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggopenstreetmap.R
\name{annotation_osm}
\alias{annotation_osm}
\title{Plot OpenStreetMap Tiles On 'ggplot2'}
\usage{
annotation_osm(obj, zoomin = +2)
}
\arguments{
\item{obj}{A `sf` object.}

\item{zoomin}{Additional zoom level.}
}
\description{
Plot OpenStreetMap Tiles On 'ggplot2'
}
\examples{
library(ggplot2)
nc <- sf::st_read(system.file("shape/nc.shp", package = "sf"), quiet = TRUE)
ggplot(nc) +
  annotation_osm(nc) +
  geom_sf(aes(fill = AREA), alpha = 0.5)
}
